TG-54: Added getProjects from TG-57 and fixed preloader and modals bugs

+++ src/Components/Projects/index.js
@@ -5,16 +5,21 @@ import Table from '../Shared/Table/Table';
 import ModalForm from '../Shared/ModalForm';
 import Form from './Form';
 import AddMember from './Form/AddMember/AddMember';
+import { useSelector, useDispatch } from 'react-redux';
+import { getProjects } from '../../redux/projects/thunks';
 
 const Projects = () => {
-  const [projects, setProjects] = useState([]);
-  const [loading, setLoading] = useState(true);
   const [showModalFormEdit, setShowModalFormEdit] = useState(false);
   const [idToEdit, setIdToEdit] = useState();
   let [value, setValue] = useState(false);
 
+  const isLoading = useSelector((state) => state.projects.isLoading);
+  const dispatch = useDispatch();
+  const projects = useSelector((state) => state.projects.list);
+
   const closeModalFormEdit = () => {
     setShowModalFormEdit(false);
+    setValue(false);
   };
 
   const openModalFormEdit = (id) => {
@@ -23,14 +28,9 @@ const Projects = () => {
   };
 
   useEffect(() => {
-    fetch(`${process.env.REACT_APP_API_URL}/api/projects`)
-      .then((response) => response.json())
-      .then((response) => {
-        setProjects(response.data);
-        setLoading(false);
-      });
+    dispatch(getProjects());
     setValue(false);
-  }, [showModalFormEdit]);
+  }, []);
 
   const deleteItem = () => {
     console.log('delete function disabled');
@@ -54,6 +54,7 @@ const Projects = () => {
           <Form
             closeModalForm={closeModalFormEdit}
             edit={true}
+            project={projects.find((project) => project._id == idToEdit)}
             itemId={idToEdit}
             functionValue={functionValue}
           />
@@ -62,30 +63,28 @@ const Projects = () => {
     );
   }
 
-  return loading ? (
-    <Preloader>
-      <p>Loading projects</p>
-    </Preloader>
+  return isLoading && !showModalFormEdit ? (
+    <Preloader />
   ) : (
     <section className={styles.container}>
       <h2 className={styles.projects}> Projects </h2>
       <Table
         data={projects}
-        headers={[
-          '_id',
-          'name',
-          'clientName',
-          'startDate',
-          'endDate',
-          'active',
-          'description',
-          'members'
+        headers={['name', 'clientName', 'startDate', 'endDate', 'active', 'description', 'members']}
+        titles={[
+          'Name',
+          'Client name',
+          'Start date',
+          'End date',
+          'Active',
+          'Description',
+          'Members'
         ]}
-        titles={['ID', 'Name', 'Clien name', 'start', 'end', 'Active', 'Description', 'Members']}
         delAction={deleteItem}
         editAction={openModalFormEdit}
       />
       {modalEdit}
+      {isLoading ? <Preloader /> : null}
       <a>
         <button className={styles.addbtn}>&#10010;</button>
       </a>
